added array_func.c to server.  autoreconf was run to regenerate src/server/Makefile.in
original job cloned for job arrays is now saved with a .TA extension in the server_priv/jobs folder.  This is deleted once the entire array is cloned. This will be used to resume the cloning process after a server restart.

for each job array we now store a file with extension .AR in server_priv/arrays
this will be used, among other things, to rebuild the array linked list on a server restart.  This is deleted when the last job belonging to an array is purged.



git-svn-id: svn://clusterresources.com/torque/trunk@1530 8f940c70-5916-0410-ac46-d1fa2fa6ea29