linked list of job arrays is recovered from files in server_priv/arrays
not fully complete,  linked list of jobs belonging to the array still needs to be recovered. I am modifying job_recov to do so.  Until then there will still be issues when a job belonging to an array is pruged after a server restart.  This should be resolved in the next day or two.



git-svn-id: svn://clusterresources.com/torque/trunk@1531 8f940c70-5916-0410-ac46-d1fa2fa6ea29