Created new definition MAXPORTLEN for the maximum string length a port value can have.
The mom and its rpp communications interface was hard coded to use pbs_rm_port for all communications. The hnodent structure has a port value that was not used and I tacked on the port from the job structure to each node requested in a job and used that value to open an new rpp communications link. I also added code to parse this new port value from the job structure.

There were problems in node_manager.c where the proper host was not getting selected for a job because of the multiple moms on a single node using the same host name. The changes made sure all selected hosts were unique and available.



git-svn-id: svn://clusterresources.com/torque/branches/2.3-multimom@2976 8f940c70-5916-0410-ac46-d1fa2fa6ea29